refactor: move regra de negócio do método cadastrar aula para a camada de serviço
